refactor: components내 icon 폴더 구조 변경
