+++ src/pages/Search/components/RecentSearchList/index.tsx
@@ -1,6 +1,6 @@
 import { IconButton } from '@mui/material';
 import { C } from './style';
-import CloseIcon from '@/components/icon/Close';
+import CloseIcon from '@/assets/svg/close.svg?react';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { deleteSearchWord, getRecentSearch } from '@/service/search';
 import { GUEST_UUID } from '@/constants/localStorage/key';
